requirements.txt
google-adk
modelcontextprotocol
gradio_client
Pillow
anyio
python-dotenv